Find pages in queue that don't have afc_state
The maintenance script that populates afc_state
looks for pages that are not in the page triage queue.

This changes includes pages that are in the page
triage queue but that don't have the afc_state
tag populated.

Bug: T203184
Change-Id: Iec255b3c4c66dba091b5e954fa7c317fafb6e20c
(cherry picked from commit a472cae878ce25715b06ad99125896958f2e021b)