Ensure a valid afc sort option is selected
This code used to try detect when an invalid sort option
was selected to change it to a valid one. It broke because
afc_state was changed from an int to a string causing it
to consider all sort options as invalid. This had consequences
when opening the filter menu and on initial page load, as reported
in the 2 bugs mentioned below.

The new approach is to use the current afc state and afc dir
to determine the only valid <option> element and select it
directly using its ID.

If the data is inconsistent or incomplete a sane default is
selected but I couldn't find a way to make this needed by
manual testing.

Selecting the <option> element by ID is needed because
jquery doesn't always select the right element with
`$('select').val('option')` when multiple options with
the same valid exist but some are hidden.

Bug: T205324
Bug: T205168
Change-Id: Ia9733476072aad0924685cca93aa97ddb0d3adfc
(cherry picked from commit 137249bb9525fc2f9f947a320282cf8d8683fc23)